+++ src/Silex/Provider/ServiceControllerServiceProvider.php
@@ -1,0 +1,32 @@
+<?php
+
+/*
+ * This file is part of the Silex framework.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Silex\Provider;
+
+use Pimple\Container;
+use Pimple\ServiceProviderInterface;
+use Silex\Application;
+use Silex\ServiceControllerResolver;
+
+class ServiceControllerServiceProvider implements ServiceProviderInterface
+{
+    public function register(Container $app)
+    {
+        $app['resolver'] = $app->share($app->extend('resolver', function ($resolver, $app) {
+            return new ServiceControllerResolver($resolver, $app);
+        }));
+    }
+
+    public function boot(Application $app)
+    {
+        // noop
+    }
+}