+++ tests/Silex/Tests/Application/UrlGeneratorTraitTest.php
@@ -18,16 +18,11 @@ use Silex\Provider\UrlGeneratorServiceProvider;
  * UrlGeneratorTrait test cases.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @requires PHP 5.4
  */
 class UrlGeneratorTraitTest extends \PHPUnit_Framework_TestCase
 {
-    public function setUp()
-    {
-        if (version_compare(phpversion(), '5.4.0', '<')) {
-            $this->markTestSkipped('PHP 5.4 is required for this test');
-        }
-    }
-
     public function testUrl()
     {
         $app = $this->createApplication();