+++ tests/Silex/Tests/RouterTest.php
@@ -27,7 +27,7 @@ class RouterTest extends \PHPUnit_Framework_TestCase
 {
     public function testMapRouting()
     {
-        $app = new Application();
+        $app = $this->createApplication();
 
         $app->match('/foo', function () {
             return 'foo';
@@ -48,7 +48,7 @@ class RouterTest extends \PHPUnit_Framework_TestCase
 
     public function testStatusCode()
     {
-        $app = new Application();
+        $app = $this->createApplication();
 
         $app->put('/created', function () {
             return new Response('', 201);
@@ -77,7 +77,7 @@ class RouterTest extends \PHPUnit_Framework_TestCase
 
     public function testRedirect()
     {
-        $app = new Application();
+        $app = $this->createApplication();
 
         $app->match('/redirect', function () {
             return new RedirectResponse('/target');
@@ -101,7 +101,7 @@ class RouterTest extends \PHPUnit_Framework_TestCase
     */
     public function testMissingRoute()
     {
-        $app = new Application();
+        $app = $this->createApplication();
         $app['exception_handler']->disable();
 
         $request = Request::create('/baz');
@@ -110,7 +110,7 @@ class RouterTest extends \PHPUnit_Framework_TestCase
 
     public function testMethodRouting()
     {
-        $app = new Application();
+        $app = $this->createApplication();
 
         $app->match('/foo', function () {
             return 'foo';
@@ -147,7 +147,7 @@ class RouterTest extends \PHPUnit_Framework_TestCase
 
     public function testRequestShouldBeStoredRegardlessOfRouting()
     {
-        $app = new Application();
+        $app = $this->createApplication();
 
         $app->get('/foo', function () use ($app) {
             return new Response($app['request']->getRequestUri());
@@ -166,7 +166,7 @@ class RouterTest extends \PHPUnit_Framework_TestCase
 
     public function testTrailingSlashBehavior()
     {
-        $app = new Application();
+        $app = $this->createApplication();
 
         $app->get('/foo/', function () use ($app) {
             return new Response('ok');
@@ -181,7 +181,7 @@ class RouterTest extends \PHPUnit_Framework_TestCase
 
     public function testRequireHttpRedirect()
     {
-        $app = new Application();
+        $app = $this->createApplication();
 
         $app->match('/secured', function () {
             return 'secured content';
@@ -195,7 +195,7 @@ class RouterTest extends \PHPUnit_Framework_TestCase
 
     public function testRequireHttpsRedirect()
     {
-        $app = new Application();
+        $app = $this->createApplication();
 
         $app->match('/secured', function () {
             return 'secured content';
@@ -209,7 +209,7 @@ class RouterTest extends \PHPUnit_Framework_TestCase
 
     public function testClassNameControllerSyntax()
     {
-        $app = new Application();
+        $app = $this->createApplication();
 
         $app->get('/foo', 'Silex\Tests\MyController::getFoo');
 
@@ -218,11 +218,16 @@ class RouterTest extends \PHPUnit_Framework_TestCase
 
     public function testClassNameControllerSyntaxWithStaticMethod()
     {
-        $app = new Application();
+        $app = $this->createApplication();
 
         $app->get('/bar', 'Silex\Tests\MyController::getBar');
 
         $this->checkRouteResponse($app, '/bar', 'bar');
+    }
+
+    protected function createApplication()
+    {
+        return new Application();
     }
 
     protected function checkRouteResponse($app, $path, $expectedContent, $method = 'get', $message = null)